% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_precip.R
\name{annual_precip}
\alias{annual_precip}
\title{annual_precip}
\usage{
annual_precip(data = "", years = 1952:2019)
}
\arguments{
\item{data}{dataframe with columns year, month, day, daily_rain}

\item{years}{the years in the dataframe you want to specify for the time period (default is 1952 to 2019)}
}
\value{
Returns a dataframe containing,
\describe{
 \item{Year}{Years specified}
 \item{Rainfall.inches}{The total precipitation (inches) for each year specified in the time period}
 }
}
\description{
Calculate annual total precipitation each year for a dataset during a specified time period.
}
\author{
Laura Ingulsrud & Keene Morrow
}
