% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_daily.R
\name{mean_daily}
\alias{mean_daily}
\title{mean_daily}
\usage{
mean_daily(data = "", start = "1900-01-01", end = "2000-01-01")
}
\arguments{
\item{data}{data frame with columns date and daily_rain}

\item{start}{the first date of the range}

\item{end}{the last date of the range}
}
\value{
Returns numeric value for the mean daily precipitation over the period between the start and end dates
}
\description{
Calculate mean daily precipitation over a specified time period. The ouput of this function returns the precipitation amount for that time period.
}
\author{
Laura Ingulsrud & Keene Morrow
}
