% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watersupply.R
\name{watersupply}
\alias{watersupply}
\title{watersupply}
\usage{
watersupply(
  data = "",
  years = 1952:2018,
  area = 100,
  total_depth = 200,
  start_depth = 100
)
}
\arguments{
\item{data}{data frame with columns year, month, day, daily_rain}

\item{years}{the years in the data frame you want to specify for the time period (default is 1952 to 2018)}

\item{area}{the surface area of the reservoir}

\item{total_depth}{the total depth of the reservoir}

\item{start_depth}{the depth at the beginning of the year range in years}
}
\value{
Returns a list containing,
\describe{
 \item{end_vol}{Volume of water in reservoir at the end of the time period (square feet)}
 \item{pct_full}{Percent of reservoir volume occupied at the end of the time period}
 \item{pct_change}{Percent change in reservoir volume occupied over the time period}
 }
}
\description{
Calculate annual total precipitation each year for UCSB during a specified time period.
}
\author{
Laura Ingulsrud & Keene Morrow
}
